Rename FormTypeField to FormTypeTitle, move fields in own folder
