+++ src/domain/pricing/modals/fields/AdditionalServicesPricingFields.tsx
@@ -3,13 +3,13 @@ import { TextInput } from 'hds-react/lib';
 import { Field } from 'formik';
 import { useTranslation } from 'react-i18next';
 
-import styles from './modals.module.scss';
-import Grid from '../../../common/grid/Grid';
-import Select from '../../../common/select/Select';
-import { formatPercentage } from '../../../common/utils/format';
-import FormTypeField from './FormTypeField';
+import styles from '../modals.module.scss';
+import Grid from '../../../../common/grid/Grid';
+import Select from '../../../../common/select/Select';
+import { formatPercentage } from '../../../../common/utils/format';
+import FormTypeTitle from '../FormTypeTitle';
 
-const AdditionalServicesFields: FunctionComponent = () => {
+const AdditionalServicesPricingFields: FunctionComponent = () => {
   const { t, i18n } = useTranslation();
 
   const serviceOptions = [
@@ -22,7 +22,7 @@ const AdditionalServicesFields: FunctionComponent = () => {
 
   return (
     <>
-      <FormTypeField
+      <FormTypeTitle
         label={t('common.terminology.dataEntry')}
         value={t('pricing.additionalServices.title')}
       />
@@ -77,4 +77,4 @@ const AdditionalServicesFields: FunctionComponent = () => {
   );
 };
 
-export default AdditionalServicesFields;
+export default AdditionalServicesPricingFields;
