+++ src/domain/pricing/modals/fields/HarborServicesPricingFields.tsx
@@ -3,10 +3,10 @@ import { TextInput } from 'hds-react/lib';
 import { Field } from 'formik';
 import { useTranslation } from 'react-i18next';
 
-import styles from './modals.module.scss';
-import Grid from '../../../common/grid/Grid';
-import Select from '../../../common/select/Select';
-import FormTypeField from './FormTypeField';
+import styles from '../modals.module.scss';
+import Grid from '../../../../common/grid/Grid';
+import Select from '../../../../common/select/Select';
+import FormTypeTitle from '../FormTypeTitle';
 
 const HarborServicesPricingFields: FunctionComponent = () => {
   const { t } = useTranslation();
@@ -24,7 +24,7 @@ const HarborServicesPricingFields: FunctionComponent = () => {
 
   return (
     <>
-      <FormTypeField
+      <FormTypeTitle
         label={t('common.terminology.dataEntry')}
         value={t('pricing.harborServices.title')}
       />
